Safari-609.1.11.1/LayoutTests/http/tests/security/same-origin-storage-blocked.html

<html>
<head>
<script>
var frames = 2;
if (window.testRunner) {
	testRunner.dumpAsText();
	testRunner.dumpChildFramesAsText();
	internals.settings.setStorageBlockingPolicy('BlockAll');
}

function decrement() {
	--frames;
	if (!frames && window.testRunner)
		internals.settings.setStorageBlockingPolicy('AllowAll');
}
</script>
</head>
<body>
<p>This iframe should return a security error:</p>
<iframe src="http://127.0.0.1:8000/security/resources/cross-origin-iframe-for-local-storage.html" onload="decrement()"></iframe>
<p>This iframe should return a security error:</p>
<iframe src="http://127.0.0.1:8000/security/resources/cross-origin-iframe-for-session-storage.html" onload="decrement()"></iframe>
</body>
</html>
