LayoutTests/http/tests/security/same-origin-storage-blocked-expected.txt

This iframe should return a security error:


This iframe should return a security error:



--------
Frame: '<!--frame1-->'
--------
SecurityError

--------
Frame: '<!--frame2-->'
--------
SecurityError
